Crate dns_parser

The network-agnostic DNS parser library

Documentation | Github | Crate

Use Builder to create a new outgoing packet.

Use Packet::parse to parse a packet into a data structure.

Re-exports

pub use rdata::RData;

Modules

rdata

Data types and methods for handling the RData field

Structs

Builder

Allows to build a DNS packet

Header

Represents parsed header of the packet

Name

The DNS name as stored in the original packet

Packet

Parsed DNS packet

Question

A parsed chunk of data in the Query section of the packet

ResourceRecord

A single DNS record

Enums

Class

The CLASS value according to RFC 1035

Error

Error parsing DNS packet

Opcode

The OPCODE value according to RFC 1035

QueryClass

The QCLASS value according to RFC 1035

QueryType

The QTYPE value according to RFC 1035

ResponseCode

The RCODE value according to RFC 1035

Type

The TYPE value according to RFC 1035