Struct dns_parser::Packet

pub struct Packet<'a> {
    pub header: Header,
    pub questions: Vec<Question<'a>>,
    pub answers: Vec<ResourceRecord<'a>>,
    pub nameservers: Vec<ResourceRecord<'a>>,
    pub additional: Vec<ResourceRecord<'a>>,
    pub opt: Option<OptRecord<'a>>,
}

Parsed DNS packet

Fields
header: Header questions: Vec<Question<'a>> answers: Vec<ResourceRecord<'a>> nameservers: Vec<ResourceRecord<'a>> additional: Vec<ResourceRecord<'a>> opt: Option<OptRecord<'a>>

Optional Pseudo-RR When present it is sent as an RR in the additional section. In this RR the class and ttl fields store max udp packet size and flags respectively. To keep ResourceRecord clean we store the OPT record here.

Methods

impl<'a> Packet<'a>

fn parse(data: &[u8]) -> Result<Packet, Error>