Struct dns_parser::ResourceRecord

pub struct ResourceRecord<'a> {
    pub name: Name<'a>,
    pub multicast_unique: bool,
    pub cls: Class,
    pub ttl: u32,
    pub data: RRData<'a>,
}

A single DNS record

We aim to provide whole range of DNS records available. But as time is limited we have some types of packets which are parsed and other provided as unparsed slice of bytes.

Fields
name: Name<'a> multicast_unique: bool

Whether or not the set of resource records is fully contained in the packet, or whether there will be more resource records in future packets. Only used for multicast DNS.

cls: Class ttl: u32 data: RRData<'a>